# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Ticaret API",
        "description": "JWT kimlik doğrulama ile E-Ticaret uygulaması için API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/cart": {
            "get": {
                "tags": [
                    "Sepet"
                ],
                "summary": "Kullanıcı sepetini getir",
                "description": "Giriş yapmış kullanıcının sepetini ve içindeki ürünleri getirir",
                "operationId": "7fdb9b40bd4ad2baf12aa15164eab788",
                "responses": {
                    "200": {
                        "description": "Sepet başarıyla getirildi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "cart": {
                                                    "type": "object"
                                                },
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "total": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz erişim - Geçersiz veya eksik JWT token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cart/add": {
            "post": {
                "tags": [
                    "Sepet"
                ],
                "summary": "Sepete ürün ekle",
                "description": "Belirtilen ürünü belirtilen miktarda sepete ekler",
                "operationId": "e940193f73a57895e17830b67c48f89c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "description": "Eklenecek ürünün ID'si",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "description": "Eklenecek ürün miktarı",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ürün sepete başarıyla eklendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ürün sepete başarıyla eklendi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz erişim - Geçersiz veya eksik JWT token"
                    },
                    "404": {
                        "description": "Ürün bulunamadı"
                    },
                    "422": {
                        "description": "Doğrulama hatası"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cart/update": {
            "put": {
                "tags": [
                    "Sepet"
                ],
                "summary": "Sepet ürün miktarını güncelle",
                "description": "Sepetteki belirli bir ürünün miktarını günceller",
                "operationId": "a97269a5c7dda30f5a76c2379a950aba",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "description": "Güncellenecek ürünün ID'si",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "description": "Yeni ürün miktarı",
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sepet başarıyla güncellendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sepet başarıyla güncellendi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz erişim - Geçersiz veya eksik JWT token"
                    },
                    "404": {
                        "description": "Ürün sepette bulunamadı"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cart/remove/{product_id}": {
            "delete": {
                "tags": [
                    "Sepet"
                ],
                "summary": "Sepetten ürün çıkar",
                "description": "Sepetten belirli bir ürünü çıkarır",
                "operationId": "759f41e0d23bd33eb8702bde3040a936",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Çıkarılacak ürünün ID'si",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ürün sepetten başarıyla çıkarıldı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ürün sepetten başarıyla çıkarıldı"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz erişim - Geçersiz veya eksik JWT token"
                    },
                    "404": {
                        "description": "Ürün sepette bulunamadı"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cart/clear": {
            "delete": {
                "tags": [
                    "Sepet"
                ],
                "summary": "Sepeti temizle",
                "description": "Sepetteki tüm ürünleri çıkarır",
                "operationId": "bd6015ca61a67dac7c1ff3bdd096bfe3",
                "responses": {
                    "200": {
                        "description": "Sepet başarıyla temizlendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sepet başarıyla temizlendi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz erişim - Geçersiz veya eksik JWT token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login veya register endpoint'inden alınan JWT token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}